backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-generativeai==0.3.2
python-dotenv==1.0.1
chromadb==0.4.22
sentence-transformers==2.3.1
